+++ lib/deprecation_toolkit/configuration.rb
@@ -12,5 +12,14 @@ module DeprecationToolkit
     config_accessor(:deprecation_path) { "test/deprecations" }
     config_accessor(:test_runner) { :minitest }
     config_accessor(:warnings_treated_as_deprecation) { [] }
+    config_accessor(:deprecation_file_path_format) do
+      proc do |test|
+        if DeprecationToolkit::Configuration.test_runner == :rspec
+          test.example_group.file_path.sub(%r{^./spec/}, "").sub(/_spec.rb$/, "")
+        else
+          test.class.name.underscore
+        end
+      end
+    end
   end
 end